Port northern lights from the Kiwicon badge.
The southern lights animation was really cool, but we can make it
even cooler now that we have color on the badge. The code is basically
the same but we add an extra x/y sin function to select the hue for
each pixel and then convert it to 24-bit RGB for rendering.